refactors ui to use ee stale endpoint
lint fixes, change endpoint location

fix failure e2e test
